- Duckcompetitoin class and its methods

+++ src/main/java/duckCasino/behaviors/fly/FlyBehavior.java
@@ -1,5 +1,5 @@
 package duckCasino.behaviors.fly;
 
 public interface FlyBehavior {
-    double fly(int time);
+    int fly(int time);
 }
